+++ app/controllers/active_authentication_controller.rb
@@ -1,0 +1,2 @@
+class ActiveAuthenticationController < ApplicationController
+end